Move new upgrade tests to old files
