+++ tests/integration/high_availability/test_upgrade_from_stable.py
@@ -1,97 +1,106 @@
 # Copyright 2023 Canonical Ltd.
 # See LICENSE file for licensing details.
 
-import asyncio
 import logging
 
+import jubilant_backports
 import pytest
-from pytest_operator.plugin import OpsTest
+from jubilant_backports import Juju
 
-from .. import juju_
-from ..helpers import get_leader_unit, get_primary_unit_wrapper, retrieve_database_variable_value
-from .high_availability_helpers import (
-    ensure_all_units_continuous_writes_incrementing,
-    relate_mysql_and_application,
+from .high_availability_helpers_new import (
+    check_mysql_units_writes_increment,
+    get_app_leader,
+    get_app_units,
+    get_mysql_primary_unit,
+    get_mysql_variable_value,
+    wait_for_apps_status,
 )
 
-logger = logging.getLogger(__name__)
-
-TIMEOUT = 20 * 60
-
 MYSQL_APP_NAME = "mysql"
-TEST_APP_NAME = "mysql-test-app"
+MYSQL_TEST_APP_NAME = "mysql-test-app"
+
+MINUTE_SECS = 60
+
+logging.getLogger("jubilant.wait").setLevel(logging.WARNING)
 
 
 @pytest.mark.abort_on_fail
-async def test_deploy_stable(ops_test: OpsTest) -> None:
-    """Simple test to ensure that the mysql and application charms get deployed."""
-    await asyncio.gather(
-        ops_test.model.deploy(
-            MYSQL_APP_NAME,
-            application_name=MYSQL_APP_NAME,
-            num_units=3,
-            channel="8.0/stable",
-            base="ubuntu@22.04",
-            config={"profile": "testing"},
-        ),
-        ops_test.model.deploy(
-            TEST_APP_NAME,
-            application_name=TEST_APP_NAME,
-            num_units=1,
-            channel="latest/edge",
-            base="ubuntu@22.04",
-        ),
+def test_deploy_stable(juju: Juju) -> None:
+    """Simple test to ensure that the MySQL and application charms get deployed."""
+    logging.info("Deploying MySQL cluster")
+    juju.deploy(
+        charm=MYSQL_APP_NAME,
+        app=MYSQL_APP_NAME,
+        base="ubuntu@22.04",
+        channel="8.0/stable",
+        config={"profile": "testing"},
+        num_units=3,
     )
-    await relate_mysql_and_application(ops_test, MYSQL_APP_NAME, TEST_APP_NAME)
-    logger.info("Wait for applications to become active")
-    await ops_test.model.wait_for_idle(
-        apps=[MYSQL_APP_NAME, TEST_APP_NAME],
-        status="active",
-        timeout=TIMEOUT,
+    juju.deploy(
+        charm=MYSQL_TEST_APP_NAME,
+        app=MYSQL_TEST_APP_NAME,
+        base="ubuntu@22.04",
+        channel="latest/edge",
+        num_units=1,
     )
-    assert len(ops_test.model.applications[MYSQL_APP_NAME].units) == 3
+
+    juju.integrate(
+        f"{MYSQL_APP_NAME}:database",
+        f"{MYSQL_TEST_APP_NAME}:database",
+    )
+
+    logging.info("Wait for applications to become active")
+    juju.wait(
+        ready=wait_for_apps_status(
+            jubilant_backports.all_active, MYSQL_APP_NAME, MYSQL_TEST_APP_NAME
+        ),
+        error=jubilant_backports.any_blocked,
+        timeout=20 * MINUTE_SECS,
+    )
 
 
 @pytest.mark.abort_on_fail
-async def test_pre_upgrade_check(ops_test: OpsTest) -> None:
+async def test_pre_upgrade_check(juju: Juju) -> None:
     """Test that the pre-upgrade-check action runs successfully."""
-    mysql_units = ops_test.model.applications[MYSQL_APP_NAME].units
+    mysql_leader = get_app_leader(juju, MYSQL_APP_NAME)
+    mysql_units = get_app_units(juju, MYSQL_APP_NAME)
 
-    logger.info("Get leader unit")
-    leader_unit = await get_leader_unit(ops_test, MYSQL_APP_NAME)
+    logging.info("Run pre-upgrade-check action")
+    task = juju.run(unit=mysql_leader, action="pre-upgrade-check")
+    task.raise_on_failure()
 
-    assert leader_unit is not None, "No leader unit found"
-    logger.info("Run pre-upgrade-check action")
-    await juju_.run_action(leader_unit, "pre-upgrade-check")
+    logging.info("Assert slow shutdown is enabled")
+    for unit_name in mysql_units:
+        value = await get_mysql_variable_value(
+            juju, MYSQL_APP_NAME, unit_name, "innodb_fast_shutdown"
+        )
+        assert value == 0
 
-    logger.info("Assert slow shutdown is enabled")
-    for unit in mysql_units:
-        value = await retrieve_database_variable_value(ops_test, unit, "innodb_fast_shutdown")
-        assert value == 0, f"innodb_fast_shutdown not 0 at {unit.name}"
-
-    primary_unit = await get_primary_unit_wrapper(ops_test, MYSQL_APP_NAME)
-
-    logger.info("Assert primary is set to leader")
-    assert await primary_unit.is_leader_from_status(), "Primary unit not set to leader"
+    logging.info("Assert primary is set to leader")
+    mysql_primary = get_mysql_primary_unit(juju, MYSQL_APP_NAME)
+    assert mysql_primary == mysql_leader, "Primary unit not set to leader"
 
 
-async def test_upgrade_from_stable(ops_test: OpsTest, charm):
-    """Test updating from stable channel."""
-    application = ops_test.model.applications[MYSQL_APP_NAME]
+@pytest.mark.abort_on_fail
+async def test_upgrade_from_stable(juju: Juju, charm: str, continuous_writes) -> None:
+    """Update the second cluster."""
+    logging.info("Ensure continuous writes are incrementing")
+    await check_mysql_units_writes_increment(juju, MYSQL_APP_NAME)
 
-    logger.info("Refresh the charm")
-    await application.refresh(path=charm)
+    logging.info("Refresh the charm")
+    juju.refresh(app=MYSQL_APP_NAME, path=charm)
 
-    logger.info("Wait for upgrade to start")
-    await ops_test.model.block_until(
-        lambda: "maintenance" in {unit.workload_status for unit in application.units},
-        timeout=TIMEOUT,
+    logging.info("Wait for upgrade to start")
+    juju.wait(
+        ready=lambda status: jubilant_backports.any_maintenance(status, MYSQL_APP_NAME),
+        timeout=10 * MINUTE_SECS,
     )
 
-    logger.info("Wait for upgrade to complete")
-    await ops_test.model.wait_for_idle(
-        apps=[MYSQL_APP_NAME], status="active", idle_period=30, timeout=TIMEOUT
+    logging.info("Wait for upgrade to complete")
+    juju.wait(
+        ready=lambda status: jubilant_backports.all_active(status, MYSQL_APP_NAME),
+        timeout=20 * MINUTE_SECS,
     )
 
-    logger.info("Ensure continuous_writes")
-    await ensure_all_units_continuous_writes_incrementing(ops_test)
+    logging.info("Ensure continuous writes are incrementing")
+    await check_mysql_units_writes_increment(juju, MYSQL_APP_NAME)
