--- tests/spread/test_upgrade_from_stable_juju_3.py/task.yaml
@@ -1,7 +1,0 @@
-summary: test_upgrade_from_stable_juju_3.py
-environment:
-  TEST_MODULE: high_availability/test_upgrade_from_stable_juju_3.py
-execute: |
-  tox run -e integration -- "tests/integration/$TEST_MODULE" --model testing --alluredir="$SPREAD_TASK/allure-results"
-artifacts:
-  - allure-results
